v7a: update barebox to v2023.09.0 -> v2023.12.0
This lets us drop two patches that fixed issues with v2023.09.0.
This update also enables some options that are useful for the hardware
we support:

  - on STM32MP1, support for the NAND controller is enabled
  - on STM32MP1, we enable HWRNG and stack protector support

The latter could be enabled universally for all platforms that are no
size constrained, but it prints an ugly error message if there's no
HWRNG as using the fixed fallback canary is insecure. We'll just
postpone enabling stack protector on other platforms until we enable
HWRNG support there.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20240112170529.1970537-5-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>