images: imx8m: move the root-a partition further back
The 896K assigned to the barebox partition are getting ever closer to
overflowing (barebox-nxp-imx8mn-evk.img is as of now 883K).
Increasing the barebox partition also requires changes to the respective
devicetrees because the location of the environment is hardcoded to
be at 896K there.

Move the root-a partition further back to free up some space so we
can add a state partition now and eventually move the environment
partition as well and then increase the barebox partition size.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
Link: https://lore.pengutronix.de/20240117133440.944800-3-l.goehrs@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>