images: imx8m: add a barebox environment partition
The barebox environment is configured in the board devicetrees to
be at 896k and 128k in size (e.g. the last 128k before the 1M mark).
This is for compatibility with the fdisk behaviour of placing the
start of the first partition at the 1M mark.

Not everyone knows this though, so it can easily happen that someone
increases the size of the barebox partition, making it overlap with
the environment "partition".

Make this less likely to happen by listing the barebox state in
the genimage config.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
Link: https://lore.pengutronix.de/20240117133440.944800-4-l.goehrs@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>