v8a: imx8m: add i.MX8MP based TQ MBa8MPxL board
The generated image can be used to boot the board via SD card.

The dev.mmc1.broken_cd barebox environment entry is required because
the board's card detect does not reliably detect an inserted USB-SD-Mux,
resulting in the board failing to boot.
This should not (negatively) affect the other v8a machines.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>
Link: https://lore.pengutronix.de/20240117133440.944800-6-l.goehrs@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>