platform-v8a: kernelconfig: cleanup
This is the result of a review of the whole kernel config.
The options have been selected by the following rules:

In general, the kernel config is optimized to keep the kernel small. So
any features that are not used at all are disabled. This includes things
like extra scheduling governors that require explicit configuration to
be used.

All drivers that are not used by the boards configured in DistroKit are
now disabled.

If there was no other strong preference then the value from the
defconfig was used.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>
Cc: Lucas Stach <l.stach@pengutronix.de>
Cc: Marc Kleine-Budde <mkl@pengutronix.de>
Link: https://lore.pengutronix.de/20240126170432.740899-1-m.olbrich@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>