v8a: imx8mq-evk: disable PCI nodes
When Linux probed the PCI on the i.MX8M-EVK, the system hangs
completely. We've seen the same effect with other i.MX8M EVKs too,
where between EVK revisions, NXP changed how the PCI PHY is clocked
and because the DT only supported one variant, probing PCI on the
other variant led to system hang while probing the PCI address space.

If this turns out to be the issue on the i.MX8M-EVK as well, we should
probably try to identify what variant we run on and then fix up the DT
in the bootloader. For now though, just take the easy route and disable
PCI, which we don't use anyway.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20240315211240.3016716-4-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>