v8a: update kernel from v6.6 → v6.8
config changes are those of olddefconfig with the notable exception of
CONFIG_BLK_DEV_WRITE_MOUNTED: This is enabled by default as not to break
some existing userspace, but we shouldn't have any of that, so disable it.
This will cause the kernel to refuse writes to mounted block devices.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20240315211240.3016716-5-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>