v8a: barebox: compress with XZ
This saves us 124K, which we dearly need, because otherwise barebox
would overlap the environment partition. The environment partition
is reserved in the genimage config, but it's used due to the description
in the barebox device tree. We may want to change both, but for now, buy
us some time by switching from LZO to XZ.

Reported-by: Michael Olbrich <mol@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20240318140633.1997386-1-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>