v7a: barebox: update to v2024.07.0
barebox v2024.07.0 was freshly relased and it includes a device tree
for the LXA MC-1 with SCMI, which will we use in a later commit,
so let's update.

Notably, this update greatly expands the number of available HWRNG
drivers, so in the future, we could configure barebox on most platforms
to enable stack protector and implement generation of the KASLR seed.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>