v7a: images: stm32mp: add support for building SCMI enabled images
The TF-A images we generate already have SCMI support. To make use of
it, we need to change used bootloader and kernel device tree to describe
access via SCMI. As the bootloader selects the device tree that matches
its own DT compatible, bootloader device tree in the FIP is the only
thing that needs to change. Let's do this by adding a @SCMI@ variable
that can be defined to be -scmi for boards that want to employ SCMI.

We add this suffix to intermediary images as well to avoid stepping on
the toes of the non-SCMI image that might be enabled in the same build.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>