v7a: barebox: am335x-mlo: document some unusual kconfig options
While we generally prefer ARM-optimized string/memory routines and
support of GPT in barebox, the MLO configuration is special:
It's size constrained by having to run from on-chip SRAM and it need
only worry about loading barebox from the same partition the MLO came
from.

Therefore leave the configuration as-is, but explain the choices made.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>