v7a: kernel: Enable I/O expander driver for STM32MP135F-DK
The i2c I/O expander MCP23S08 provides GPIOs to reset various components
(ethernet phy, display bridge, camera). The I/O expander is configured
as built-in as it is needed for nfs-root operation (once the kernel
contains the needed adaptions for the stmmac to support the stm32mp13x).

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>