mipsel: sync kernelconfig
Part of the changes stem from the toolchain update (commit
282b31d940f79ed01352, 2024-04-22, Ladislav Michl: "mipsel: update to
OSELAS.Toolchain 2023.07.1"), other parts in the kernel-malta config
were just out of sync with the main kernelconfig. Use default values for
all prompted options.

The changes in the PAHOLE version stem from the fact that our kernel
rules are outdated, which we'll fix in later commits. Nevertheless, keep
the kernelconfigs in sync so as not to break bisectability.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>