mips, mipsel: remove forked kernel rule
Our forked kernel rule is outdated, but it is only a stripped-down
version of the upstream kernel rule that does not actually build a
kernel, but is used to provide a REF_CONFIG for other kernel recipes.
Remove the forked rule to make future maintainance easier, and set the
platformconfig options and the KERNEL_MAKE_ENV suitably so that the
upstream kernel rule does the same as before. Also run 'oldconfig' to
update the kernelconfig and pick up the changes from the upstream kernel
rule.

The changes in the PAHOLE version in the kernelconfig diffs stem from
the fact that the kernel rules are outdated, which we'll fix in later
commits. Nevertheless, keep the kernelconfigs in sync so as not to break
bisectability.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>