mips, mipsel: migrate barebox-common recipe to barebox
With a bit of hackery in barebox.noexec.make, we can use the upstream
barebox package to provide a REF_CONFIG for the other barebox packages.
The slight disadvantage is that its targetinstall stage creates an empty
barebox-default-environment file in the image folder, but the big
advantage is that we can save maintaining a custom rule.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>