mipsel: qemu-run: enable emulation of little endian barebox
qemu-system-mipsel unlike GXemul requires that the 32-bit words inside
the barebox binary are swapped when not using the default big-endian
mode. Recent barebox versions generate an additional swapped image
that can be used, so let's import the barebox patch that enables it.

We can drop this patch again, when update to a barebox version that
includes the stdout-path changes, probably v2024.10.0 or later.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>