v7a: tf-a: update v2.10 -> v2.12
Like done for v8a, let's bring TF-A for v7a up to date as well.
This needs a slight adjustment beyond updating the platformconfig:

The build system has been reworked in the new TF-A versions and
specifying our own CC or AS variables now breaks the build[1].

This has been fixed in PTXdist already, but as we have a second TF-A
recipe for the STM32MP13, we must apply the change there as well.

[1]: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/27163
[2]: https://git.pengutronix.de/cgit/ptxdist/commit/rules/tf-a.make?id=341428f78

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>