v7a: barebox: am335x-mlo: disable MTD
We currently configure the MLO image to support SPI-NOR and NAND flash,
but we only generate SD/eMMC images for am335x platforms.

The MLO first stage always boots from the same boot medium it came from
and thus the MTD support is unused. barebox v2025.03.0 got big enough
that it breaks our config, thus lets disable MTD for now.

There are already fixes to get the size down queued for the next
release, so this commit can be reverted in future.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>